third_party/blink/web_tests/platform/generic/fast/dom/Window/console-functions-expected.txt

CONSOLE MESSAGE: console.log() called.
CONSOLE MESSAGE: console.info() called.
CONSOLE WARNING: console.warn() called.
CONSOLE ERROR: console.error() called.
Tests window.console functions. To make sure this test passed you must check the console for the following 4 messages:

console.log() called.
console.info() called.
console.warn() called.
console.error() called.
